State in the art of MQL/C++ compatibility.

+++ SymbolInfo.static.h
@@ -138,7 +138,8 @@ class SymbolInfoStatic {
    * See: http://forum.mql4.com/30672
    */
   static unsigned int GetPointsPerPip(string _symbol) {
-    return (unsigned int)pow(10, SymbolInfoStatic::GetDigits(_symbol) - SymbolInfoStatic::GetPipDigits(_symbol));
+    return (unsigned int)pow((unsigned int)10,
+                             SymbolInfoStatic::GetDigits(_symbol) - SymbolInfoStatic::GetPipDigits(_symbol));
   }
 
   /**
@@ -264,7 +265,7 @@ class SymbolInfoStatic {
    * Get real spread based on the ask and bid price (in points).
    */
   static unsigned int GetRealSpread(double _bid, double _ask, unsigned int _digits) {
-    return (unsigned int)round((_ask - _bid) * pow(10, _digits));
+    return (unsigned int)round((_ask - _bid) * pow((unsigned int)10, _digits));
   }
   static unsigned int GetRealSpread(string _symbol) {
     return GetRealSpread(SymbolInfoStatic::GetBid(_symbol), SymbolInfoStatic::GetAsk(_symbol),
